Extracting and Visualizing Stock Data
Project Overview:
Assuming the role of Data Scientist / Data Analyst. Extracted financial data like historical share price and quarterly revenue reportings from various sources using Python libraries and webscraping on popular stocks. After collecting this data, visualized it in a dashboard to identify patterns or trends. The stocks we will work with are Tesla and GameStop.

This notebook will be used for the Peer Reviewed Assignment of the "Python Project for Data Science" course.